mlops_platform/data_manager_service/requirements.txt
flask
flask_cors
python-configuration[yaml]
smart_open[s3]
jsonlines
tinydb
matplotlib
soundfile
docker
pyyaml
pydub